requirements.txt
altair==5.1.2
annotated-types==0.6.0
anyio==4.0.0
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.3
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
deprecation==2.1.0
extra-streamlit-components==0.1.60
gitdb==4.0.11
GitPython==3.1.40
gotrue==1.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.26.1
packaging==23.2
pandas==2.1.1
Pillow==10.1.0
postgrest==0.11.0
protobuf==4.24.4
pyarrow==13.0.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
realtime==1.0.0
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
st-supabase-connection==1.2.0
storage3==0.6.1
streamlit==1.28.0
streamlit-authenticator==0.2.3
streamlit-survey==0.1.0
StrEnum==0.4.15
supabase==1.2.0
supafunc==0.2.3
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
watchdog==3.0.0
websockets==10.4
zipp==3.17.0
